Try fix ytdl not working

+++ main.py
@@ -18,7 +18,8 @@ def main() -> None:
     from neonbot import bot
     from neonbot.utils.constants import PLAYER_CACHE_DIR, YOUTUBE_DOWNLOADS_DIR
 
-    shutil.rmtree(YOUTUBE_DOWNLOADS_DIR, ignore_errors=True)
+    if env.bool('YTDL_AUTO_CLEAR_DOWNLOADS', default=False):
+        shutil.rmtree(YOUTUBE_DOWNLOADS_DIR, ignore_errors=True)
     os.makedirs(YOUTUBE_DOWNLOADS_DIR, exist_ok=True)
     os.makedirs(PLAYER_CACHE_DIR, exist_ok=True)
